requirements.txt
numpy==1.19.5
Cython==0.29.22
cmake==3.22.2
dlib==19.23.0
onnxruntime==1.10.0
insightface==0.5
face-recognition==1.3.0
deepface==0.0.72
python-dotenv==0.19.2
kafka-python==2.0.2
sentry-sdk==1.5.1
